Fix previous commit

+++ minobot/src/bot.rs
@@ -163,7 +163,7 @@ impl Node {
         }
     }
     fn expand<E: Evaluator>(&mut self, data: &BotData<E>) -> ((i32, i32), u32) {
-        if self.depth >= data.queue.len() {
+        if self.depth as usize >= data.queue.len() {
             self.finished = true;
             return ((std::i32::MIN, 0), 0);
         }
